+++ app/__init__.py
@@ -1,1 +1,1 @@
-"""Docstring."""
+""" #todo написать докстринг как в README"""
