+++ app/storage.py
@@ -5,7 +5,7 @@ from typing import List, Optional
 
 import aioredis
 
-from app.bike_model import Bike, SubscriptionBikeFamily
+from app.models import Bike, SubscriptionBikeFamily
 from app.settings import app_settings
 
 BIKE_KEY = 'canyon-notifier:bike:{0}'
@@ -123,7 +123,6 @@ async def save_new_available_bikes(available_bikes_list: List[Bike]) -> None:
 
 async def get_available_bike_list() -> List[Bike]:
     """Get bikes, now available in the store, from database. Return list of it."""
-    # todo test
     output: List[Bike] = []
     bike_id_list = await db_pool.smembers(SUBSCRIPTION_AVAILABLE_KEY)
 
@@ -139,7 +138,6 @@ async def get_available_bike_list() -> List[Bike]:
 
 async def delete_available_bike_list(bike_id_list: Optional[List[str]] = None) -> None:
     """Delete items from the list of available bikes."""
-    # todo test
     if bike_id_list is None:
         await db_pool.delete(SUBSCRIPTION_AVAILABLE_KEY)
     else:
