+++ tests/test_subscription_notifier/test_get_notification_bikes.py
@@ -1,5 +1,5 @@
-from app.bike_model import Bike, SubscriptionBikeFamily, Match
-from app.subscription_notifier import send_subscription_message, get_notification_bikes, main
+from app.models import Bike, SubscriptionBikeFamily, Match
+from app.subscription_notifier import get_notification_bikes
 
 
 def test_get_notification_bikes_happy_path():
@@ -79,35 +79,3 @@ def test_get_notification_bikes_happy_path():
         )
     ]
 
-
-async def test_send_subscription_message_smoke():
-    # todo проверку на конкретное сообщение
-    subscription_to_send = Match(
-        Bike(
-            id='spectral_125_cf_9',
-            title='Spectral 125 CF 9',
-            link='https://www.canyon.com/en-cz/mountain-bikes/trail-bikes/spectral-125/cf/spectral-125-cf-9/3179.html?dwvar_3179_pv_rahmenfarbe=SR',
-            family='Spectral',
-            model='125 CF 9',
-        ),
-        SubscriptionBikeFamily(
-            subscribe_id=12345678,
-            chat_id=915745042,
-            bike_family='Spectral',
-        ),
-    )
-    expected_message = '\n'.join((
-        'Spectral 125 CF 9 is available in the stock!',
-        'https://www.canyon.com/en-cz/mountain-bikes/trail-bikes/spectral-125/cf/spectral-125-cf-9/3179.html?dwvar_3179_pv_rahmenfarbe=SR',
-    ))
-
-    res = await send_subscription_message(subscription_to_send)
-
-    assert res is True
-
-
-async def test_main_smoke():
-    res = await main(throttling_time=5.0, amount_of_iterations=2)
-
-    assert True
-    assert res == 2
