+++ tests/tests_catalog_updater/test_parse_canyon_catalog.py
@@ -1,7 +1,7 @@
 from lxml import etree
 
 from app.catalog_updater import _parse_canyon_catalog
-from app.bike_model import Bike
+from app.models import Bike
 
 
 def test_parse_canyon_catalog_one_bike():
