+++ tests/tests_storage/test_insert_actual_catalog.py
@@ -1,6 +1,6 @@
 import pytest
 
-from app.bike_model import Bike
+from app.models import Bike
 from app.storage import insert_actual_catalog
